[material-ui-nextjs] Fix build size (#40436)
