+++ packages/mui-material-nextjs/src/v13-pagesRouter/index.ts
@@ -1,1 +1,2 @@
-export * from './pagesRouterV13';
+export * from './pagesRouterV13Document';
+export * from './pagesRouterV13App';
