+++ packages/mui-material-nextjs/src/v13-pagesRouter/pagesRouterV13Document.tsx
@@ -1,40 +1,10 @@
 import * as React from 'react';
 import { AppType } from 'next/app';
-import { CacheProvider, EmotionCache } from '@emotion/react';
-import createCache from '@emotion/cache';
+import { EmotionCache } from '@emotion/react';
 import createEmotionServer from '@emotion/server/create-instance';
 import Document, { DocumentContext, DocumentInitialProps } from 'next/document';
-
-const isBrowser = typeof document !== 'undefined';
-
-// On the client side, Create a meta tag at the top of the <head> and set it as insertionPoint.
-// This assures that MUI styles are loaded first.
-// It allows developers to easily override MUI styles with other styling solutions, like CSS modules.
-function createEmotionCache() {
-  let insertionPoint;
-
-  if (isBrowser) {
-    const emotionInsertionPoint = document.querySelector<HTMLMetaElement>(
-      'meta[name="emotion-insertion-point"]',
-    );
-    insertionPoint = emotionInsertionPoint ?? undefined;
-  }
-
-  return createCache({ key: 'mui', insertionPoint });
-}
-
-interface EmotionCacheProviderProps {
-  emotionCache?: EmotionCache;
-}
-
-const defaultEmotionCache = createEmotionCache();
-
-export function AppCacheProvider({
-  emotionCache = defaultEmotionCache,
-  children,
-}: React.PropsWithChildren<EmotionCacheProviderProps>) {
-  return <CacheProvider value={emotionCache}>{children}</CacheProvider>;
-}
+import { EmotionCacheProviderProps } from './pagesRouterV13App';
+import createEmotionCache from './createCache';
 
 interface Plugin {
   enhanceApp: (
